+++ frontend/src/routeTree.gen.ts
@@ -80,6 +80,7 @@ import { Route as ServerRequiredAuthenticatedCGuildIdSettingsInitiativesRouteImp
 import { Route as ServerRequiredAuthenticatedCGuildIdSettingsTrashRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/settings/trash'
 import { Route as ServerRequiredAuthenticatedCGuildIdSettingsUsersRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/settings/users'
 import { Route as ServerRequiredAuthenticatedCGuildIdTagsTagIdRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/tags_.$tagId'
+import { Route as ServerRequiredAuthenticatedCGuildIdUUserIdRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/u.$userId'
 import { Route as ServerRequiredAuthenticatedCGuildIdCalendarsCalendarIdIndexRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/calendars/$calendarId/index'
 import { Route as ServerRequiredAuthenticatedCGuildIdCalendarsCalendarIdSettingsRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/calendars/$calendarId/settings'
 import { Route as ServerRequiredAuthenticatedCGuildIdGoRefTypeRefIdRouteImport } from './routes/_serverRequired/_authenticated/c/$guildId/go/$refType/$refId'
@@ -557,6 +558,12 @@ const ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute =
   ServerRequiredAuthenticatedCGuildIdTagsTagIdRouteImport.update({
     id: '/tags_/$tagId',
     path: '/tags/$tagId',
+    getParentRoute: () => ServerRequiredAuthenticatedCGuildIdRoute,
+  } as any)
+const ServerRequiredAuthenticatedCGuildIdUUserIdRoute =
+  ServerRequiredAuthenticatedCGuildIdUUserIdRouteImport.update({
+    id: '/u/$userId',
+    path: '/u/$userId',
     getParentRoute: () => ServerRequiredAuthenticatedCGuildIdRoute,
   } as any)
 const ServerRequiredAuthenticatedCGuildIdCalendarsCalendarIdIndexRoute =
@@ -1151,6 +1158,7 @@ export interface FileRoutesByFullPath {
   '/c/$guildId/settings/trash': typeof ServerRequiredAuthenticatedCGuildIdSettingsTrashRoute
   '/c/$guildId/settings/users': typeof ServerRequiredAuthenticatedCGuildIdSettingsUsersRoute
   '/c/$guildId/tags/$tagId': typeof ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute
+  '/c/$guildId/u/$userId': typeof ServerRequiredAuthenticatedCGuildIdUUserIdRoute
   '/c/$guildId/calendars/': typeof ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute
   '/c/$guildId/i/': typeof ServerRequiredAuthenticatedCGuildIdIIndexRoute
   '/c/$guildId/settings/': typeof ServerRequiredAuthenticatedCGuildIdSettingsIndexRoute
@@ -1276,6 +1284,7 @@ export interface FileRoutesByTo {
   '/c/$guildId/settings/trash': typeof ServerRequiredAuthenticatedCGuildIdSettingsTrashRoute
   '/c/$guildId/settings/users': typeof ServerRequiredAuthenticatedCGuildIdSettingsUsersRoute
   '/c/$guildId/tags/$tagId': typeof ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute
+  '/c/$guildId/u/$userId': typeof ServerRequiredAuthenticatedCGuildIdUUserIdRoute
   '/c/$guildId/calendars': typeof ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute
   '/c/$guildId/i': typeof ServerRequiredAuthenticatedCGuildIdIIndexRoute
   '/c/$guildId/settings': typeof ServerRequiredAuthenticatedCGuildIdSettingsIndexRoute
@@ -1402,6 +1411,7 @@ export interface FileRoutesById {
   '/_serverRequired/_authenticated/c/$guildId/settings/trash': typeof ServerRequiredAuthenticatedCGuildIdSettingsTrashRoute
   '/_serverRequired/_authenticated/c/$guildId/settings/users': typeof ServerRequiredAuthenticatedCGuildIdSettingsUsersRoute
   '/_serverRequired/_authenticated/c/$guildId/tags_/$tagId': typeof ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute
+  '/_serverRequired/_authenticated/c/$guildId/u/$userId': typeof ServerRequiredAuthenticatedCGuildIdUUserIdRoute
   '/_serverRequired/_authenticated/c/$guildId/calendars/': typeof ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute
   '/_serverRequired/_authenticated/c/$guildId/i/': typeof ServerRequiredAuthenticatedCGuildIdIIndexRoute
   '/_serverRequired/_authenticated/c/$guildId/settings/': typeof ServerRequiredAuthenticatedCGuildIdSettingsIndexRoute
@@ -1535,6 +1545,7 @@ export interface FileRouteTypes {
     | '/c/$guildId/settings/trash'
     | '/c/$guildId/settings/users'
     | '/c/$guildId/tags/$tagId'
+    | '/c/$guildId/u/$userId'
     | '/c/$guildId/calendars/'
     | '/c/$guildId/i/'
     | '/c/$guildId/settings/'
@@ -1660,6 +1671,7 @@ export interface FileRouteTypes {
     | '/c/$guildId/settings/trash'
     | '/c/$guildId/settings/users'
     | '/c/$guildId/tags/$tagId'
+    | '/c/$guildId/u/$userId'
     | '/c/$guildId/calendars'
     | '/c/$guildId/i'
     | '/c/$guildId/settings'
@@ -1785,6 +1797,7 @@ export interface FileRouteTypes {
     | '/_serverRequired/_authenticated/c/$guildId/settings/trash'
     | '/_serverRequired/_authenticated/c/$guildId/settings/users'
     | '/_serverRequired/_authenticated/c/$guildId/tags_/$tagId'
+    | '/_serverRequired/_authenticated/c/$guildId/u/$userId'
     | '/_serverRequired/_authenticated/c/$guildId/calendars/'
     | '/_serverRequired/_authenticated/c/$guildId/i/'
     | '/_serverRequired/_authenticated/c/$guildId/settings/'
@@ -2353,6 +2366,13 @@ declare module '@tanstack/react-router' {
       path: '/tags/$tagId'
       fullPath: '/c/$guildId/tags/$tagId'
       preLoaderRoute: typeof ServerRequiredAuthenticatedCGuildIdTagsTagIdRouteImport
+      parentRoute: typeof ServerRequiredAuthenticatedCGuildIdRoute
+    }
+    '/_serverRequired/_authenticated/c/$guildId/u/$userId': {
+      id: '/_serverRequired/_authenticated/c/$guildId/u/$userId'
+      path: '/u/$userId'
+      fullPath: '/c/$guildId/u/$userId'
+      preLoaderRoute: typeof ServerRequiredAuthenticatedCGuildIdUUserIdRouteImport
       parentRoute: typeof ServerRequiredAuthenticatedCGuildIdRoute
     }
     '/_serverRequired/_authenticated/c/$guildId/calendars/$calendarId/': {
@@ -3219,6 +3239,7 @@ interface ServerRequiredAuthenticatedCGuildIdRouteChildren {
   ServerRequiredAuthenticatedCGuildIdIInitiativeIdRoute: typeof ServerRequiredAuthenticatedCGuildIdIInitiativeIdRouteWithChildren
   ServerRequiredAuthenticatedCGuildIdMarketplacePublicIdRoute: typeof ServerRequiredAuthenticatedCGuildIdMarketplacePublicIdRoute
   ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute: typeof ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute
+  ServerRequiredAuthenticatedCGuildIdUUserIdRoute: typeof ServerRequiredAuthenticatedCGuildIdUUserIdRoute
   ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute: typeof ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute
   ServerRequiredAuthenticatedCGuildIdIIndexRoute: typeof ServerRequiredAuthenticatedCGuildIdIIndexRoute
   ServerRequiredAuthenticatedCGuildIdCalendarsCalendarIdSettingsRoute: typeof ServerRequiredAuthenticatedCGuildIdCalendarsCalendarIdSettingsRouteWithChildren
@@ -3246,6 +3267,8 @@ const ServerRequiredAuthenticatedCGuildIdRouteChildren: ServerRequiredAuthentica
       ServerRequiredAuthenticatedCGuildIdMarketplacePublicIdRoute,
     ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute:
       ServerRequiredAuthenticatedCGuildIdTagsTagIdRoute,
+    ServerRequiredAuthenticatedCGuildIdUUserIdRoute:
+      ServerRequiredAuthenticatedCGuildIdUUserIdRoute,
     ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute:
       ServerRequiredAuthenticatedCGuildIdCalendarsIndexRoute,
     ServerRequiredAuthenticatedCGuildIdIIndexRoute: